Split main_test.go into separate files
