+++ test/e2e/main/spec_update_test.go
@@ -1,0 +1,117 @@
+package main
+
+import (
+	"context"
+	"testing"
+
+	"github.com/iancoleman/strcase"
+	ispnv1 "github.com/infinispan/infinispan-operator/api/v1"
+	tutils "github.com/infinispan/infinispan-operator/test/e2e/utils"
+	appsv1 "k8s.io/api/apps/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/wait"
+)
+
+// Test if spec.container.cpu update is handled
+func TestContainerCPUUpdateWithTwoReplicas(t *testing.T) {
+	defer testKube.CleanNamespaceAndLogOnPanic(t, tutils.Namespace)
+
+	var modifier = func(ispn *ispnv1.Infinispan) {
+		ispn.Spec.Container.CPU = "900m:550m"
+	}
+	var verifier = func(ispn *ispnv1.Infinispan, ss *appsv1.StatefulSet) {
+		limit := resource.MustParse("900m")
+		request := resource.MustParse("550m")
+		if limit.Cmp(ss.Spec.Template.Spec.Containers[0].Resources.Limits["cpu"]) != 0 ||
+			request.Cmp(ss.Spec.Template.Spec.Containers[0].Resources.Requests["cpu"]) != 0 {
+			panic("CPU field not updated")
+		}
+	}
+	spec := tutils.MinimalSpec
+	spec.Name = strcase.ToKebab(t.Name())
+	spec.Labels = map[string]string{"test-name": t.Name()}
+	genericTestForContainerUpdated(spec, modifier, verifier)
+}
+
+// Test if spec.container.memory update is handled
+func TestContainerMemoryUpdate(t *testing.T) {
+	t.Parallel()
+	defer testKube.CleanNamespaceAndLogOnPanic(t, tutils.Namespace)
+
+	var modifier = func(ispn *ispnv1.Infinispan) {
+		ispn.Spec.Container.Memory = "512Mi:256Mi"
+	}
+	var verifier = func(ispn *ispnv1.Infinispan, ss *appsv1.StatefulSet) {
+		limit := resource.MustParse("512Mi")
+		request := resource.MustParse("256Mi")
+		if limit.Cmp(ss.Spec.Template.Spec.Containers[0].Resources.Limits["memory"]) != 0 ||
+			request.Cmp(ss.Spec.Template.Spec.Containers[0].Resources.Requests["memory"]) != 0 {
+			panic("Memory field not updated")
+		}
+	}
+	spec := tutils.DefaultSpec(t, testKube)
+	genericTestForContainerUpdated(*spec, modifier, verifier)
+}
+
+func TestContainerJavaOptsUpdate(t *testing.T) {
+	t.Parallel()
+	defer testKube.CleanNamespaceAndLogOnPanic(t, tutils.Namespace)
+
+	var modifier = func(ispn *ispnv1.Infinispan) {
+		ispn.Spec.Container.ExtraJvmOpts = "-XX:NativeMemoryTracking=summary"
+	}
+	var verifier = func(ispn *ispnv1.Infinispan, ss *appsv1.StatefulSet) {
+		env := ss.Spec.Template.Spec.Containers[0].Env
+		for _, value := range env {
+			if value.Name == "JAVA_OPTIONS" {
+				if value.Value != "-XX:NativeMemoryTracking=summary" {
+					panic("JAVA_OPTIONS not updated")
+				} else {
+					return
+				}
+			}
+		}
+		panic("JAVA_OPTIONS not updated")
+	}
+	spec := tutils.DefaultSpec(t, testKube)
+	genericTestForContainerUpdated(*spec, modifier, verifier)
+}
+
+func genericTestForContainerUpdated(ispn ispnv1.Infinispan, modifier func(*ispnv1.Infinispan), verifier func(*ispnv1.Infinispan, *appsv1.StatefulSet)) {
+	testKube.CreateInfinispan(&ispn, tutils.Namespace)
+	testKube.WaitForInfinispanPods(int(ispn.Spec.Replicas), tutils.SinglePodTimeout, ispn.Name, tutils.Namespace)
+	testKube.WaitForInfinispanCondition(ispn.Name, ispn.Namespace, ispnv1.ConditionWellFormed)
+
+	verifyStatefulSetUpdate(ispn, modifier, verifier)
+}
+
+func verifyStatefulSetUpdate(ispn ispnv1.Infinispan, modifier func(*ispnv1.Infinispan), verifier func(*ispnv1.Infinispan, *appsv1.StatefulSet)) {
+	// Get the associate StatefulSet
+	ss := appsv1.StatefulSet{}
+	// Get the current generation
+	tutils.ExpectNoError(testKube.Kubernetes.Client.Get(context.TODO(), types.NamespacedName{Namespace: ispn.Namespace, Name: ispn.GetStatefulSetName()}, &ss))
+	generation := ss.Status.ObservedGeneration
+
+	tutils.ExpectNoError(testKube.UpdateInfinispan(&ispn, func() {
+		modifier(&ispn)
+	}))
+
+	// Wait for a new generation to appear
+	err := wait.Poll(tutils.DefaultPollPeriod, tutils.SinglePodTimeout, func() (done bool, err error) {
+		tutils.ExpectNoError(testKube.Kubernetes.Client.Get(context.TODO(), types.NamespacedName{Namespace: ispn.Namespace, Name: ispn.Name}, &ss))
+		return ss.Status.ObservedGeneration >= generation+1, nil
+	})
+	tutils.ExpectNoError(err)
+
+	// Wait that current and update revisions match
+	// this ensures that the rolling upgrade completes
+	err = wait.Poll(tutils.DefaultPollPeriod, tutils.SinglePodTimeout, func() (done bool, err error) {
+		tutils.ExpectNoError(testKube.Kubernetes.Client.Get(context.TODO(), types.NamespacedName{Namespace: ispn.Namespace, Name: ispn.Name}, &ss))
+		return ss.Status.CurrentRevision == ss.Status.UpdateRevision, nil
+	})
+	tutils.ExpectNoError(err)
+
+	// Check that the update has been propagated
+	verifier(&ispn, &ss)
+}
